server/requirements.txt
antlr4-python3-runtime==4.8
appdirs==1.4.4
asgiref==3.5.0
astropy==4.2
audioread==2.1.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.0
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.13
django-allauth==0.50.0
django-cors-headers==3.11.0
django-crontab==0.7.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
fonttools==4.33.3
hydra-core==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.2.3
jamo==0.4.1
jiwer==2.3.0
joblib==1.1.0
kiwisolver==1.4.2
librosa==0.9.1
llvmlite==0.38.0
matplotlib==3.5.1
mysqlclient==2.1.0
numba==0.55.1
numpy==1.21.6
oauthlib==3.2.0
omegaconf==2.1.2
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
Pillow==9.1.0
pooch==1.6.0
pycparser==2.21
pyerfa==2.0.0.1
PyJWT==1.7.1
pyparsing==3.0.8
python-dateutil==2.8.2
python-Levenshtein==0.12.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.96
six==1.16.0
SoundFile==0.10.3.post1
sqlparse==0.4.2
threadpoolctl==3.1.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
typing-extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
zipp==3.8.0
